Made a geojson file for each county
counties.geojson was too large and cumbersome.  I found better geojson
at https://github.com/johan/world.geo.json/ and decided to store each
county as a separate file in a “geojson” directory.  Adding support for
a new county is as easy as dropping a new file in said directory.